projects/assignment 1/requirements.txt
numpy
pandas
pillow
matplotlib
opencv-python
scipy
